+++ Tourist.Server/Forms/EntityForm.cs
@@ -89,7 +89,8 @@ namespace Tourist.Server.Forms
 
 					AddEntityToRepository( buffer );
 
-					//MetroMessageBox.Show( this, "Do you like this metro message box?", "Metro Title", MessageBoxButtons.OK, MessageBoxIcon.Information );
+					MetroMessageBox.Show( this, "Entity Added With Sucess !!!", "Metro Title", MessageBoxButtons.OK, MessageBoxIcon.Information );
+
 				}
 				else
 				{
@@ -192,7 +193,10 @@ namespace Tourist.Server.Forms
 		{
 			var removeIndex = e.RowIndex;
 
-			var aRow = EntityDataGrid.Rows[ e.RowIndex ];
+			DialogResult dialog = MetroMessageBox.Show( this, "Are you sure you want to remove the entity at row number " + ( e.RowIndex + 1 ) + " ?", "Metro Title", MessageBoxButtons.YesNo, MessageBoxIcon.Information );
+
+			if ( dialog == DialogResult.No )
+				return;
 
 			repository.RemoveEntity( removeIndex );
